trading_engine/requirements.txt
ccxt
pandas
numpy
sqlalchemy
python-dotenv
schedule
textblob
ta
SQLAlchemy
psycopg2-binary
uvicorn
